requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0
sqlalchemy==2.0.23
instagrapi==2.0.0
playwright==1.40.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
pandas==2.1.4
openpyxl==3.1.2
urllib3==2.1.0
Pillow==10.1.0
fake-useragent==1.4.0